# backend/contract/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "Code-first Filebonsai API. The PostgreSQL profile authenticates the local owner and derives catalog scope from server-side membership.",
    "title" : "Filebonsai Catalog",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "http://127.0.0.1:8080"
  } ],
  "security" : [ {
    "sessionCookie" : [ ]
  } ],
  "paths" : {
...
    "/api/v1/auth/login" : {
      "post" : {
        "operationId" : "login",
        "parameters" : [ {
          "in" : "header",
          "name" : "X-CSRF-TOKEN",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AccessLoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccessSessionResponse"
                }
              }
            },
            "description" : "Authenticated session"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "VALIDATION_FAILED"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INVALID_CREDENTIALS"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "CSRF_INVALID"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "RATE_LIMITED",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds until another login attempt is accepted",
                "schema" : {
                  "minimum" : 1,
                  "type" : "integer",
                  "format" : "int64"
                },
                "style" : "simple"
              }
            }
          }
        },
        "security" : [ ],
        "summary" : "Authenticate the local owner and rotate the session",
        "tags" : [ "Access" ]
      }
    },
    "/api/v1/auth/logout" : {
      "post" : {
        "operationId" : "logout",
        "parameters" : [ {
          "in" : "header",
          "name" : "X-CSRF-TOKEN",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Session invalidated"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "AUTH_REQUIRED"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "CSRF_INVALID"
          }
        },
        "summary" : "Invalidate the current session",
        "tags" : [ "Access" ]
      }
    },
    "/api/v1/auth/me" : {
      "get" : {
        "operationId" : "getCurrentSession",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccessSessionResponse"
                }
              }
            },
            "description" : "Authenticated session"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "AUTH_REQUIRED"
          }
        },
        "summary" : "Read the authenticated local owner",
        "tags" : [ "Access" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "AccessLoginRequest" : {
        "required" : [ "password" ],
        "type" : "object",
        "properties" : {
          "password" : {
            "maxLength" : 1024,
            "minLength" : 1,
            "pattern" : "^[\\s\\S]*\\S[\\s\\S]*$",
            "type" : "string",
            "description" : "Contains at least one non-whitespace character",
            "writeOnly" : true
          }
        }
      },
      "AccessSessionResponse" : {
        "required" : [ "expiresAt", "principalId" ],
        "type" : "object",
        "properties" : {
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "principalId" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      },
      "ApiErrorResponse" : {
        "required" : [ "code", "fieldErrors", "message", "requestId", "status" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "Open string, not an enum. Known codes are documented per operation. Fall back to status and message for unknown codes."
          },
          "fieldErrors" : {
            "type" : "array",
            "description" : "Empty when no individual field is responsible",
            "items" : {
              "$ref" : "#/components/schemas/FieldErrorResponse"
            }
          },
          "message" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "FieldErrorResponse" : {
        "required" : [ "code", "field", "message" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "Open string code; clients must allow unknown codes"
          },
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "description" : "Empty when no individual field is responsible"
      },
...
    },
    "securitySchemes" : {
      "sessionCookie" : {
        "in" : "cookie",
        "name" : "FILEBONSAI_SESSION",
        "type" : "apiKey"
      }
    }
  }
}